C++ committee really wastes time on useless shit like this
source/gameshared/gs_misc.cpp:54:19: error: arithmetic between different enumeration types ('GadgetType' and 'WeaponType') is deprecated [-Werror,-Wdeprecated-enum-enum-conversion]

+++ source/gameshared/gs_misc.cpp
@@ -47,15 +47,15 @@ void GS_TouchPushTrigger( const gs_state_t * gs, SyncPlayerState * playerState, 
 }
 
 DamageType::DamageType( WeaponType weapon ) {
-	encoded = weapon;
+	encoded = u8( weapon );
 }
 
 DamageType::DamageType( GadgetType gadget ) {
-	encoded = gadget + Weapon_Count;
+	encoded = u8( gadget ) + u8( Weapon_Count );
 }
 
 DamageType::DamageType( WorldDamage world ) {
-	encoded = world + Weapon_Count + Gadget_Count;
+	encoded = u8( world ) + u8( Weapon_Count ) + u8( Gadget_Count );
 }
 
 bool operator==( DamageType a, DamageType b ) {
